learnRL/rlCookbook/requirements.txt
ale-py==0.7.4
atari-py==0.2.9
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
box2d-py==2.3.5
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.2
cloudpickle==1.2.2
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
future==0.18.2
gym-notices==0.0.6
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
importlib-resources==5.6.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
numpy==1.21.4
opencv-python==4.5.5.64
Pillow==9.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pygame==2.1.0
pyglet==1.3.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
zipp==3.8.0
